Removed circular depenency
