+++ packages/core/src/client.test.ts
@@ -1,5 +1,4 @@
 import { Bot, Bundle, Identifier, Patient, SearchParameter, StructureDefinition } from '@medplum/fhirtypes';
-import { MockAsyncClientStorage } from '@medplum/mock';
 import { randomUUID, webcrypto } from 'crypto';
 import PdfPrinter from 'pdfmake';
 import type { CustomTableLayout, TDocumentDefinitions, TFontDictionary } from 'pdfmake/interfaces';
@@ -10,7 +9,6 @@ import {
   FetchLike,
   InviteRequest,
   MedplumClient,
-  MedplumClientOptions,
   NewPatientRequest,
   NewProjectRequest,
   NewUserRequest,
@@ -18,6 +16,7 @@ import {
 import { mockFetch } from './client-test-utils';
 import { ContentType } from './contenttype';
 import { OperationOutcomeError, notFound, unauthorized } from './outcomes';
+import { MockAsyncClientStorage } from './storage';
 import { getDataType, isDataTypeLoaded, isProfileLoaded } from './typeschema/types';
 import { ProfileResource, createReference } from './utils';
 
@@ -2695,7 +2694,7 @@ describe('Passed in async-backed `ClientStorage`', () => {
   test('MedplumClient resolves initialized after storage is initialized', async () => {
     const fetch = mockFetch(200, { success: true });
     const storage = new MockAsyncClientStorage();
-    const medplum = new MedplumClient({ fetch, storage } as unknown as MedplumClientOptions);
+    const medplum = new MedplumClient({ fetch, storage });
     expect(storage.isInitialized).toEqual(false);
     expect(medplum.isInitialized).toEqual(false);
     storage.setInitialized();
